update to docker-17.06.0-ce

+++ provision-docker.ps1
@@ -1,10 +1,17 @@
 # see https://docs.microsoft.com/en-us/virtualization/windowscontainers/manage-docker/configure-docker-daemon
-# see https://github.com/moby/moby/releases/tag/v17.05.0-ce-rc3
+# see https://docs.docker.com/engine/installation/linux/docker-ce/binaries/#install-server-and-client-binaries-on-windows
+# see https://github.com/docker/docker-ce/releases/tag/v17.06.0-ce
 
 # download install the docker binaries.
-$archiveName = 'docker-17.05.0-ce-rc3.zip'
+$archiveName = 'docker-17.06.0-ce.zip'
+$archiveUrl = "https://download.docker.com/win/static/stable/x86_64/$archiveName"
+$archiveHash = '3d27360a11a3a627aac9c6d73eb32d4a9b6dcca6bcb4b2c7a5fcd9d2e0ec6c82'
 $archivePath = "$env:TEMP\$archiveName"
-Invoke-WebRequest https://test.docker.com/builds/Windows/x86_64/$archiveName -UseBasicParsing -OutFile $archivePath
+Invoke-WebRequest $archiveUrl -UseBasicParsing -OutFile $archivePath
+$archiveActualHash = (Get-FileHash $archivePath -Algorithm SHA256).Hash
+if ($archiveActualHash -ne $archiveHash) {
+    throw "the $archiveUrl file hash $archiveActualHash does not match the expected $archiveHash"
+}
 Expand-Archive $archivePath -DestinationPath $env:ProgramFiles
 Remove-Item $archivePath
 
@@ -18,49 +25,30 @@ $env:PATH += ";$env:ProgramFiles\docker"
 
 # install the docker service and configure it to always restart on failure.
 dockerd --register-service
-sc.exe failure docker reset= 0 actions= restart/1000/restart/1000/restart/1000
+sc.exe failure docker reset= 0 actions= restart/1000
 
 # configure docker through a configuration file.
 # see https://docs.docker.com/engine/reference/commandline/dockerd/#windows-configuration-file
 $config = @{
     'debug' = $false
     'labels' = @('os=windows')
-    'hosts' = @('tcp://0.0.0.0:2375', 'npipe:////./pipe/docker_engine')
+    'hosts' = @(
+        'tcp://0.0.0.0:2375',
+        'npipe:////./pipe/docker_engine'
+    )
 }
-mkdir -Force "$env:ProgramData\docker\config"
+mkdir -Force "$env:ProgramData\docker\config" | Out-Null
 Set-Content -Encoding ascii "$env:ProgramData\docker\config\daemon.json" ($config | ConvertTo-Json)
 
-# start docker.
+Write-Host 'Starting docker...'
 Start-Service docker
 
-function docker {
-    docker.exe @Args
-    if ($LASTEXITCODE) {
-        throw "$(@('docker')+$Args | ConvertTo-Json -Compress) failed with exit code $LASTEXITCODE"
-    }
+function Write-Title($title) {
+    Write-Output "#`n# $title`n#"
 }
 
-Write-Host 'Pulling the microsoft/nanoserver container image...'
-docker pull microsoft/nanoserver
-
-Write-Host 'Pulling and running microsoft/dotnet-samples:dotnetapp-nanoserver...'
-docker run --rm microsoft/dotnet-samples:dotnetapp-nanoserver
-
-Write-Host '# docker version'
+Write-Title 'docker version'
 docker version
 
-Write-Host '# docker info'
+Write-Title 'docker info'
 docker info
-
-Write-Host '# docker network ls'
-docker network ls
-
-Write-Host '# docker images'
-docker images
-
-Write-Host '# docker disk space statistics'
-docker system df -v
-
-choco install -y docker-compose
-Write-Host '# docker-compose version'
-docker-compose version
